requirements.txt
openai==1.9.0
python-dotenv==0.21.0